build: fix install path for dbus service
On NixOS all packages are installed into their own immutable prefix.
Because of this dbus.get_pkgconfig_variable will return a
path from within dbus's prefix and we cannot write to it.
By using define_variable we can replace the datadir to
be from the paths from meson. This should have no affect
on elementaryOS.